+++ tests/test_admin.py
@@ -82,6 +82,34 @@ class TestAdmin(TestCase, WagtailTestUtils):
         assert new_page.child_of(TrashCanPage.objects.first())
         assert TrashCan.objects.count() == 1
 
+    def test_bulk_delete_sets_and_unsets_slug(self):
+        from wagtail_trash.wagtail_hooks import urlconf_time
+
+        root_page = Page.objects.get(url_path="/")
+
+        new_page1 = Page(title="new page 1")
+
+        p1 = root_page.add_child(instance=new_page1)
+
+        p1_slug = p1.slug
+
+        with self.register_hook("before_bulk_action", trash_bulk_delete):
+            delete_url = reverse(
+                "wagtail_bulk_action", args=("wagtailcore", "page", "delete")
+            )
+            id_query = "?id=" + "&id=".join([str(p1.id)])
+            self.client.post(delete_url + id_query)
+
+        p1.refresh_from_db()
+        self.assertEqual(f"trash-{p1.pk}-{p1_slug}", p1.slug)
+
+        with self.register_hook("register_admin_urls", urlconf_time):
+            restore_url = reverse("wagtail_trash_restore", args=(p1.pk,))
+            self.client.get(restore_url)
+
+        p1.refresh_from_db()
+        self.assertEqual(p1_slug, p1.slug)
+
     def test_bulk_delete_sends_to_trash_can(self):
         root_page = Page.objects.get(url_path="/")
 
@@ -127,6 +155,27 @@ class TestAdmin(TestCase, WagtailTestUtils):
             self.client.post(delete_url)
             assert not Page.objects.filter(title="delete page")
             assert TrashCan.objects.count() == 0
+
+    def test_removing_two_pages_with_same_slug(self):
+        root_page = Page.objects.get(url_path="/")
+
+        cat = Page(title="category 1", slug="category-1")
+        cat_instance = root_page.add_child(instance=cat)
+        cat_2 = Page(title="category 2", slug="category-2")
+        cat_instance_2 = root_page.add_child(instance=cat_2)
+
+        delete = Page(title="delete", slug="delete")
+        delete_instance = cat_instance.add_child(instance=delete)
+        delete_2 = Page(title="delete", slug="delete")
+        delete_instance_2 = cat_instance_2.add_child(instance=delete_2)
+
+        with self.register_hook("before_delete_page", trash_delete):
+            delete_url = reverse("wagtailadmin_pages:delete", args=(delete.id,))
+            self.client.post(delete_url)
+            delete_url = reverse("wagtailadmin_pages:delete", args=(delete_2.id,))
+            self.client.post(delete_url)
+
+            self.assertEqual(TrashCan.objects.count(), 2)
 
     def test_removing_page_unpublishes_all_sub_pages(self):
         root_page = Page.objects.get(url_path="/")